Add custom lifecycle to Maven plugin
This makes available a custom lifecycle via the Maven plugin, consisting
of the following phases:

1. lk-pre-eval
2. lk-eval
3. lk-post-eval

This allows evaluation to be run in evaluation projects without firing
the rest of the Maven build machinery.  Eventually, we may also want to
implement an additional packaging type and do some work with lifecycle
mappings.