Create simplified evaluation runner (refs #115)
This commit has the initial code for a new evaluation runner along with integration tests for it in the lenskit-eval project.  The lenskit-eval-ant-plugin is now outdated and will be removed or replaced shortly.

I also deleted the old data-based tests in lenskit-knn; they will replaced with new ones using the upcoming Maven plugin.