Fixed for eclipse defaults.
Fixes so there are reasonable defaults for the eclipse install, so pure eclipse users can run integration tests without having to configure anything, as long as they create a "data" directory and unzip a download of "ml-100k" into it.  There is a bug in the archetype tests in this version.