Allow file parameters to be bound to strings
This bolts on some machinery to allow File parameters to be bound to String values, making it easier to write configuration files.

Implemented by augmenting the LensKit binding implementation with a type coercion function, applied to instance bindings.  `set` on a File parameter applies a string-to-file coercion, allowing the string parameters.