Fix crossfold output routing
This adds output file support to the crossfold spec so that we can get that information before running the crossfolder.

This means that the crossfold spec knows a lot about how the crossfolder produces output. This should be capitalized on in the crossfolder itself so that we don't duplicate that logic.